linked projects controller
