+++ app.js
@@ -2,8 +2,9 @@ import mongoose from "mongoose";
 import express from "express";
 import cors from "cors";
 import session from "express-session";
-import UserController from "./users/users-controller.js";
-import ApiController from "./apis/api-controller.js";
+import UsersController from "./users/users-controller.js";
+import ApisController from "./apis/apis-controller.js";
+import ProjectsController from "./projects/projects-controller.js"
 
 const CONNECTION_STRING = "mongodb+srv://krugert:ZIFQQxbMj7zC7Pmp@neu-collab-0.1oel9h6.mongodb.net/";
 
@@ -18,8 +19,9 @@ app.use(
 
 app.use(express.json());
 
-UserController(app);
-ApiController(app);
+UsersController(app);
+ProjectsController(app);
+ApisController(app);
 
 
 app.listen(4000, () => {
