+++ projects/projects-dao.js
@@ -1,4 +1,4 @@
-import projectsModel from "./projects-model";
+import projectsModel from "./projects-model.js";
 
 export const findAll = () => {
     return projectsModel.find();
@@ -27,5 +27,3 @@ export const deleteById = async (id) => {
     projectsModel.deleteOne({_id});
     return {status: 'ok'};
 }
-
-export default ProjectsDAO;
