+++ projects/projects-model.js
@@ -1,5 +1,5 @@
 import mongoose, { mongo } from "mongoose";
-import projectsSchema from "./projects-schema";
+import projectsSchema from "./projects-schema.js";
 
 const projectsModel = mongoose.model("Projects", projectsSchema);
 export default projectsModel;
